+++ nix/foldl.nix
@@ -1,7 +1,7 @@
 { mkDerivation, base, bytestring, comonad, containers
 , contravariant, criterion, hashable, mwc-random, primitive
-, profunctors, semigroupoids, semigroups, stdenv, text
-, transformers, unordered-containers, vector, vector-builder
+, profunctors, safe-exceptions, semigroupoids, semigroups, stdenv
+, text, transformers, unordered-containers, vector, vector-builder
 }:
 mkDerivation {
   pname = "foldl";
@@ -9,8 +9,9 @@ mkDerivation {
   src = ./..;
   libraryHaskellDepends = [
     base bytestring comonad containers contravariant hashable
-    mwc-random primitive profunctors semigroupoids semigroups text
-    transformers unordered-containers vector vector-builder
+    mwc-random primitive profunctors safe-exceptions semigroupoids
+    semigroups text transformers unordered-containers vector
+    vector-builder
   ];
   benchmarkHaskellDepends = [ base criterion ];
   description = "Composable, streaming, and efficient left folds";