+++ src/Control/Foldl.hs
@@ -39,7 +39,9 @@
 {-# LANGUAGE ExistentialQuantification #-}
 {-# LANGUAGE FlexibleContexts          #-}
 {-# LANGUAGE RankNTypes                #-}
+{-# LANGUAGE ScopedTypeVariables       #-}
 {-# LANGUAGE Trustworthy               #-}
+{-# LANGUAGE TypeApplications          #-}
 
 module Control.Foldl (
     -- * Fold Types
@@ -130,6 +132,10 @@ module Control.Foldl (
     , groupBy
     , either
     , eitherM
+    , exHalt_
+    , exHalt
+    , exSkip_
+    , exSkip
 
     -- * Re-exports
     -- $reexports
@@ -141,6 +147,7 @@ module Control.Foldl (
 import Control.Foldl.Optics (_Left, _Right)
 import Control.Applicative
 import Control.Foldl.Internal (Maybe'(..), lazy, Either'(..), Pair(..), hush)
+import Control.Exception.Safe (Exception, MonadCatch, SomeException, try)
 import Control.Monad ((<=<))
 import Control.Monad.Primitive (PrimMonad, RealWorld)
 import Control.Comonad
@@ -1347,3 +1354,134 @@ eitherM l r = (,) <$> handlesM _Left l <*> handlesM _Right r
 
     @Data.Vector.Generic@ re-exports the 'Vector' type class
 -}
+
+{-| When 'IO' is the monad of a monadic fold, each step in the evaluation of the
+fold might throw an exception. 'exHalt_' gives you the ability to recover the
+fold's result from the steps preceding the failure.
+
+>>> import Control.Exception
+>>> f x = if x < 10 then return x else throw Overflow
+>>> xs = [1, 2, 500, 4] :: [Integer]
+
+Since @f 500@ produces an exception, the following fold fails:
+
+>>> fold1 = premapM f (generalize list)
+>>> foldM fold1 xs
+*** Exception: arithmetic overflow
+
+By applying 'untilFirstException', we can produce a new fold that returns
+the intermediate result at the point where the exception occurs.
+
+>>> fold2 = exHalt_ fold1
+>>> foldM fold2 xs
+[1,2]
+
+The fold internally uses 'try' from the @safe-exceptions@ package and will not
+catch async exceptions.
+-}
+
+exHalt_ :: forall m a b. (Monad m, MonadCatch m) => FoldM m a b -> FoldM m a b
+exHalt_ f = snd <$> exHalt @SomeException f
+
+{-| Like 'exHalt_', but also returns the exception thrown (if any) and provides
+the ability to specify what type of exception to catch. Any other type of
+exception will still cause the entire fold's evaluation to fail.
+
+>>> import Control.Exception
+>>> f x = if x < 10 then return x else throw Overflow
+>>> xs = [1, 2, 500, 4] :: [Integer]
+>>> fold1 = premapM f (generalize list)
+
+>>> :set -XTypeApplications
+>>> fold2 = exHalt @ArithException fold1
+
+>>> foldM fold2 xs
+(Just arithmetic overflow,[1,2])
+-}
+
+exHalt :: forall e m a b. (Exception e, Monad m, MonadCatch m)
+    => FoldM m a b -> FoldM m a (Maybe e, b)
+exHalt (FoldM step begin done) = FoldM step' begin' done'
+  where
+    begin' =
+      do
+        y <- begin
+        return (Nothing, y)
+
+    step' x'@(Just _, _) _ = return x'
+    step' (Nothing, x1) a =
+      do
+        x2Either <- try (step x1 a)
+        case x2Either of
+            Left e   -> return (Just e, x1)
+            Right x2 -> return (Nothing, x2)
+
+    done' (eMaybe, x) =
+      do
+        b <- done x
+        return (eMaybe, b)
+
+{-| When 'IO' is the monad of a monadic fold, each step in the evaluation of the
+fold might throw an exception. 'exSkip_' gives you the ability to perform only
+steps that succeed, ignoring any steps that failed.
+
+>>> import Control.Exception
+>>> f x = if x < 10 then return x else throw Overflow
+>>> xs = [1, 2, 500, 4] :: [Integer]
+
+Since @f 500@ produces an exception, the following fold fails:
+
+>>> fold1 = premapM f (generalize list)
+>>> foldM fold1 xs
+*** Exception: arithmetic overflow
+
+By applying 'exSkip_', we can produce a new fold that produces a result from all
+steps that /don't/ fail:
+
+>>> fold2 = exSkip_ fold1
+>>> foldM fold2 xs
+[1,2,4]
+
+The fold internally uses 'try' from the @safe-exceptions@ package and will not
+catch async exceptions.
+-}
+
+exSkip_ :: forall m a b. (Monad m, MonadCatch m) => FoldM m a b -> FoldM m a b
+exSkip_ f = snd <$> exSkip @SomeException f
+
+{-| Like 'exSkip_', but also returns the exceptions thrown (if any) and provides
+the ability to specify what type of exception to catch. Any other type of
+exception will still cause the entire fold's evaluation to fail.
+
+>>> import Control.Exception
+>>> f x = if x < 10 then return x else throw Overflow
+>>> xs = [1, 2, 500, 4] :: [Integer]
+>>> fold1 = premapM f (generalize list)
+
+>>> :set -XTypeApplications
+>>> fold2 = exSkip @ArithException fold1
+
+>>> foldM fold2 xs
+([arithmetic overflow],[1,2,4])
+-}
+
+exSkip :: forall e m a b. (Exception e, Monad m, MonadCatch m)
+    => FoldM m a b -> FoldM m a ([e], b)
+exSkip (FoldM step begin done) = FoldM step' begin' done'
+  where
+    begin' =
+      do
+        y <- begin
+        return (id, y)
+
+    step' (es, x1) a =
+      do
+        x2Either <- try (step x1 a)
+        case x2Either of
+            Left e   -> return (es . (e :), x1)
+            Right x2 -> return (es, x2)
+
+    done' (es, x) =
+      do
+        b <- done x
+        return (es [], b)